refact/moved queryKind files to queryKind/dml since they were all dml query kinds
