+++ src/queryKinds/dml/insert.ts
@@ -1,7 +1,7 @@
-import CteMaker, { Cte } from "../cteMaker.js";
-import SqlEscaper from "../sqlEscaper.js";
-import ColumnValue from "../types/ColumnValue.js";
-import QueryKind from "../types/QueryKind.js";
+import CteMaker, { Cte } from "../../cteMaker.js";
+import SqlEscaper from "../../sqlEscaper.js";
+import ColumnValue from "../../types/ColumnValue.js";
+import QueryKind from "../../types/QueryKind.js";
 import QueryDefinition from "./query.js";
 import SelectQuery from "./select.js";
