+++ src/queryKinds/dml/query.ts
@@ -1,16 +1,16 @@
 import { ValidatorOptions } from "class-validator";
-import deepEqual from "../deepEqual.js";
-import { getClassValidator, getZod } from "../getOptionalPackages.js";
+import deepEqual from "../../deepEqual.js";
+import { getClassValidator, getZod } from "../../getOptionalPackages.js";
 // Import types only since they are used for type checking only
 // and zod is optional peer dependency
 import type z from "zod";
 import type { ZodObject } from "zod";
-import sqlFlavor from "../types/sqlFlavor.js";
-import CteMaker from "../cteMaker.js";
-import Statement from "../statementMaker.js";
-import QueryKind from "../types/QueryKind.js";
-import Join, { isJoinTable } from "../types/Join.js";
-import SqlEscaper from "../sqlEscaper.js";
+import sqlFlavor from "../../types/sqlFlavor.js";
+import CteMaker from "../../cteMaker.js";
+import Statement from "../../statementMaker.js";
+import QueryKind from "../../types/QueryKind.js";
+import Join, { isJoinTable } from "../../types/Join.js";
+import SqlEscaper from "../../sqlEscaper.js";
 
 /**
   * An array of function names that can be used to execute SQL queries.
